requirements.txt
discord.py==2.0.1
hikari==2.0.0.dev111
hikari_lightbulb==2.2.4
hikari-lightbulb
nextcord==2.2.0
